backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
huggingface-hub
supabase